Simplify FolderHandler
