+++ pkg/grafana/folder-handler.go
@@ -4,8 +4,15 @@ import (
 	"fmt"
 	"path/filepath"
 
+	"encoding/json"
+	"errors"
+
 	"github.com/grafana/grizzly/pkg/grizzly"
 	"github.com/grafana/tanka/pkg/kubernetes/manifest"
+
+	gclient "github.com/grafana/grafana-openapi-client-go/client"
+	"github.com/grafana/grafana-openapi-client-go/client/folders"
+	"github.com/grafana/grafana-openapi-client-go/models"
 )
 
 // FolderHandler is a Grizzly Handler for Grafana dashboard folders
@@ -87,7 +94,7 @@ func (h *FolderHandler) GetUID(resource grizzly.Resource) (string, error) {
 
 // GetByUID retrieves JSON for a resource from an endpoint, by UID
 func (h *FolderHandler) GetByUID(UID string) (*grizzly.Resource, error) {
-	resource, err := getRemoteFolder(h.Provider.client, UID)
+	resource, err := h.getRemoteFolder(UID)
 	if err != nil {
 		return nil, fmt.Errorf("Error retrieving dashboard folder %s: %w", UID, err)
 	}
@@ -97,20 +104,142 @@ func (h *FolderHandler) GetByUID(UID string) (*grizzly.Resource, error) {
 
 // GetRemote retrieves a folder as a resource
 func (h *FolderHandler) GetRemote(resource grizzly.Resource) (*grizzly.Resource, error) {
-	return getRemoteFolder(h.Provider.client, resource.Name())
+	return h.getRemoteFolder(resource.Name())
 }
 
 // ListRemote retrieves as list of UIDs of all remote resources
 func (h *FolderHandler) ListRemote() ([]string, error) {
-	return getRemoteFolderList(h.Provider.client)
+	return h.getRemoteFolderList()
 }
 
 // Add pushes a new folder to Grafana via the API
 func (h *FolderHandler) Add(resource grizzly.Resource) error {
-	return postFolder(h.Provider.client, resource)
+	return h.postFolder(resource)
 }
 
 // Update pushes a folder to Grafana via the API
 func (h *FolderHandler) Update(existing, resource grizzly.Resource) error {
-	return putFolder(h.Provider.client, resource)
+	return h.putFolder(resource)
+}
+
+// getRemoteFolder retrieves a folder object from Grafana
+func (h *FolderHandler) getRemoteFolder(uid string) (*grizzly.Resource, error) {
+	var folder *models.Folder
+	if uid == "General" || uid == "general" {
+		folder = &models.Folder{
+			ID:    0,
+			UID:   uid,
+			Title: "General",
+			// URL: ??
+		}
+	} else {
+		params := folders.NewGetFolderByUIDParams().WithFolderUID(uid)
+		folderOk, err := h.Provider.client.Folders.GetFolderByUID(params, nil)
+		if err != nil {
+			var gErrNotFound *folders.GetFolderByUIDNotFound
+			var gErrForbidden *folders.GetFolderByUIDForbidden
+			if errors.As(err, &gErrNotFound) || errors.As(err, &gErrForbidden) {
+				return nil, fmt.Errorf("Couldn't fetch folder '%s' from remote: %w", uid, grizzly.ErrNotFound)
+			}
+			return nil, err
+		}
+		folder = folderOk.GetPayload()
+	}
+
+	// TODO: Turn spec into a real models.Folder object
+	spec, err := structToMap(folder)
+	if err != nil {
+		return nil, err
+	}
+
+	resource := grizzly.NewResource(h.APIVersion(), h.Kind(), uid, spec)
+	return &resource, nil
+}
+
+func (h *FolderHandler) getRemoteFolderList() ([]string, error) {
+	var (
+		limit       = int64(1000)
+		page  int64 = 0
+		uids  []string
+	)
+	params := folders.NewGetFoldersParams().WithLimit(&limit)
+	for {
+		page++
+		params.SetPage(&page)
+
+		foldersOk, err := h.Provider.client.Folders.GetFolders(params, nil)
+		if err != nil {
+			return nil, err
+		}
+
+		for _, folder := range foldersOk.GetPayload() {
+			uids = append(uids, folder.UID)
+		}
+		if int64(len(foldersOk.GetPayload())) < *params.Limit {
+			return uids, nil
+		}
+	}
+}
+
+func (h *FolderHandler) postFolder(resource grizzly.Resource) error {
+	name := resource.Name()
+	if name == "General" || name == "general" {
+		return nil
+	}
+
+	// TODO: Turn spec into a real models.Folder object
+	data, err := json.Marshal(resource.Spec())
+	if err != nil {
+		return err
+	}
+
+	var folder models.Folder
+	err = json.Unmarshal(data, &folder)
+	if err != nil {
+		return err
+	}
+	if folder.Title == "" {
+		return fmt.Errorf("missing title in folder spec")
+	}
+
+	body := models.CreateFolderCommand{
+		Title: folder.Title,
+		UID:   folder.UID,
+	}
+	params := folders.NewCreateFolderParams().WithBody(&body)
+	_, err = h.Provider.client.Folders.CreateFolder(params, nil)
+	return err
+}
+
+func (h *FolderHandler) putFolder(resource grizzly.Resource) error {
+	// TODO: Turn spec into a real models.Folder object
+	data, err := json.Marshal(resource.Spec())
+	if err != nil {
+		return err
+	}
+
+	var folder models.Folder
+	err = json.Unmarshal(data, &folder)
+	if err != nil {
+		return err
+	}
+	if folder.Title == "" {
+		return fmt.Errorf("missing title in folder spec")
+	}
+
+	body := models.UpdateFolderCommand{
+		Title: folder.Title,
+	}
+	params := folders.NewUpdateFolderParams().WithBody(&body)
+	_, err = h.Provider.client.Folders.UpdateFolder(params, nil)
+	return err
+}
+
+var getFolderById = func(client *gclient.GrafanaHTTPAPI, folderId int64) (*models.Folder, error) {
+	params := folders.NewGetFolderByIDParams().WithFolderID(folderId)
+	folderOk, err := client.Folders.GetFolderByID(params, nil)
+	if err != nil {
+		return nil, err
+	}
+	return folderOk.GetPayload(), nil
 }
